# ExtendedSchedule/domain/values/common.go
package values

type AudienceInfo struct {
	ID   string `json:"id"`
	Name string `json:"name"`
}

type GroupInfo struct {
	ID   string `json:"id"`
	Name string `json:"name"`
}

type TeacherInfo struct {
	ID   string `json:"id"`
	Name string `json:"name"`
}

type FacultyInfo struct {
	ID   string `json:"id"`
	Name string `json:"name"`
}