# ExtendedSchedule/services/tsuschedule/openapi.yaml
openapi: 3.0.0
info:
  title: TSU Schedule
  description: TSU schedule public API
  version: 0.1.0
servers:
  - url: https://intime.tsu.ru/api/web/v1
    description: Web API path

paths:
  /schedule/group:
    get:
      summary: Schedule by group
      tags:
        - ScheduleAPI
      parameters:
        - name: id
          in: query
          required: true
          description: groupd identifier
          schema:
            type: string
        - name: dateFrom
          in: query
          required: true
          description: "date to start from"
          schema:
            type: string
            example: "2022-04-17"
        - name: dateTo
          in: query
          required: true
          description: "date to end"
          schema:
            type: string
            example: "2022-04-24"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                title: "Schedule list by day"
                type: array
                items:
                  $ref: "#/components/schemas/DaySchedule"
  /faculties:
    get:
      summary: Get all faculties
      tags:
        - ScheduleAPI
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                title: "List of all faculties"
                type: array
                items:
                  $ref: "#/components/schemas/FacultyInfo"

  /faculties/{facultyID}/groups:
    get:
      summary: Get all groups for specific faculty
      tags:
        - ScheduleAPI
      parameters:
        - name: facultyID
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                title: "List of groups"
                type: array
                items:
                  $ref: "#/components/schemas/GroupInfo"

components:
  schemas:
    DaySchedule:
      type: object
      title: "Day schedule"
      required:
        - date
        - lessons
      properties:
        date:
          type: "string"
          example: "2022-04-19"
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
    Lesson:
      type: object
      title: "Lesson"
      required:
        - type
        - lessonNumber
      properties:
        type:
          type: "string"
          example: "LESSON"
        lessonNumber:
          type: integer
          example: 1
        lessonType:
          type: "string"
          example: "PRACTICE"
        id:
          type: "string"
        title:
          type: string
          example: "Computer Science"
        audience:
          $ref: "#/components/schemas/AudienceInfo"
        groups:
          type: array
          items:
            $ref: "#/components/schemas/GroupInfo"
        professor:
          $ref: "#/components/schemas/TeacherInfo"
    GroupInfo:
      type: object
      title: "Group Info"
      required:
        - id
        - name
      properties:
        id:
          type: string
          example: "group-identifier"
        name:
          type: string
          example: "931901"
    TeacherInfo:
      type: object
      title: "Teacher Info"
      properties:
        id:
          type: string
          example: "teacher-identifier"
        fullName:
          type: string
          example: "Teacher Name"
    AudienceInfo:
      type: object
      title: "Audience Info"
      required:
        - id
        - name
      properties:
        id:
          type: string
          nullable: true
          example: "audience-identifier"
        name:
          type: string
          example: "104"
    FacultyInfo:
      type: object
      title: "Faculty Info"
      required:
        - id
        - name
      properties:
        id:
          type: string
          example: "faculty-identifier"
        name:
          type: string
          example: "Институт Прикладной Математики и Компьютерных Наук"